fix(login-auth): fix feature login
