+++ package.json
@@ -29,12 +29,12 @@
     "fs": "^0.0.1-security",
     "lucide-react": "^0.441.0",
     "mongoose": "^8.8.1",
-    "next": "14.2.11",
+    "next": "^14.2.11",
     "next-auth": "^4.24.10",
     "next-themes": "^0.3.0",
     "papaparse": "^5.4.1",
     "react": "^18.3.1",
-    "react-dom": "^18",
+    "react-dom": "^18.3.1",
     "react-hook-form": "^7.53.1",
     "react-icons": "^5.3.0",
     "tailwind-merge": "^2.5.4",
